Less strict code of the TinylogLoggingProvider #168.
Move methods of common use out of ConfigurationParser and
TinylogLoggingProvider to allow user providers with less code
duplication as discussed in issue #168.
