+++ tinylog-impl/src/test/java/org/tinylog/core/ConfigurationParserTest.java
@@ -13,9 +13,6 @@
 
 package org.tinylog.core;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 
@@ -27,15 +24,8 @@ import org.powermock.reflect.Whitebox;
 import org.tinylog.Level;
 import org.tinylog.configuration.Configuration;
 import org.tinylog.rules.SystemStreamCollector;
-import org.tinylog.throwable.ThrowableFilter;
-import org.tinylog.util.ConfigurationStoreWriter;
-import org.tinylog.writers.ConsoleWriter;
-import org.tinylog.writers.FileWriter;
-import org.tinylog.writers.Writer;
 
-import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
-import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
 
@@ -279,238 +269,6 @@ public final class ConfigurationParserTest {
 
 		boolean enabled = ConfigurationParser.isAutoShutdownEnabled();
 		assertThat(enabled).isTrue();
-	}
-
-	/**
-	 * Verifies that a default writer will be created for all tags and severity levels, if logging is enabled but no
-	 * writer explicitly defined.
-	 */
-	@Test
-	public void defaultWriter() {
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.TRACE, false);
-
-		assertThat(writers).hasSize(2).allSatisfy(element ->
-			assertThat(element).hasSize(5).allSatisfy(collection ->
-				assertThat(collection).hasSize(1).allSatisfy(writer -> assertThat(writer).isNotNull())
-			)
-		);
-	}
-
-	/**
-	 * Verifies that no writer will be created, if logging is disabled.
-	 */
-	@Test
-	public void noWriters() {
-		Configuration.set("writer", "console");
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.OFF, false);
-
-		assertThat(writers).hasSize(2).allSatisfy(element ->
-			assertThat(element).hasSize(5).allSatisfy(collection -> assertThat(collection).isEmpty())
-		);
-	}
-
-	/**
-	 * Verifies that a single tagged writer will be created and assigned correctly.
-	 */
-	@Test
-	public void singleTaggedWriter() {
-		Configuration.set("writer", "console");
-		Configuration.set("writer.tag", "SYSTEM");
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(singletonList("SYSTEM"), Level.TRACE, false);
-
-		assertThat(writers)
-			.hasSize(3)
-			.allSatisfy(element -> assertThat(element).hasSize(5));
-
-		assertThat(writers[0]).allSatisfy(collection -> assertThat(collection).isEmpty());
-		assertThat(writers[1]).allSatisfy(collection ->
-			assertThat(collection).hasSize(1).allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class))
-		);
-		assertThat(writers[2]).allSatisfy(collection -> assertThat(collection).isEmpty());
-	}
-
-	/**
-	 * Verifies that a single explicitly untagged writer will be created and assigned correctly.
-	 */
-	@Test
-	public void singleUntaggedWriter() {
-		Configuration.set("writer", "console");
-		Configuration.set("writer.tag", "-");
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.TRACE, false);
-
-		assertThat(writers).hasSize(2);
-		assertThat(writers[0]).allSatisfy(collection ->
-			assertThat(collection).hasSize(1).allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class))
-		);
-		assertThat(writers[1]).allSatisfy(collection -> assertThat(collection).isEmpty());
-	}
-
-	/**
-	 * Verifies that a single tagged writer with multiple tags will be created and assigned correctly.
-	 */
-	@Test
-	public void singleMultipleTaggedWriter() {
-		Configuration.set("writer", "console");
-		Configuration.set("writer.tag", " system , , backup  , test, , "); // Test also unusual tag entries
-		List<String> tags = ConfigurationParser.getTags();
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(tags, Level.TRACE, false);
-
-		assertThat(writers)
-			.hasSize(5)
-			.allSatisfy(element -> assertThat(element).hasSize(5));
-
-		assertThat(writers[0]).allSatisfy(collection -> assertThat(collection).isEmpty());
-		assertThat(writers[1]).allSatisfy(collection -> {
-			assertThat(collection).hasSize(1).allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-		});
-		assertThat(writers[2]).allSatisfy(collection -> {
-			assertThat(collection).hasSize(1).allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-		});
-		assertThat(writers[3]).allSatisfy(collection -> {
-			assertThat(collection).hasSize(1).allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-		});
-		assertThat(writers[4]).allSatisfy(collection -> assertThat(collection).isEmpty());
-	}
-	
-	/**
-	 * Verifies that two tagged writers will be created and assigned correctly.
-	 *
-	 * @throws IOException
-	 *             Failed to create temporary log file
-	 */
-	@Test
-	public void multipleTaggedWriter() throws IOException {
-		Configuration.set("writer1", "console");
-		Configuration.set("writer1.tag", "SYSTEM");
-
-		Configuration.set("writer2", "file");
-		Configuration.set("writer2.tag", "SYSTEM");
-		Configuration.set("writer2.file", File.createTempFile("tinylog", ".log").getAbsolutePath());
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(singletonList("SYSTEM"), Level.TRACE, false);
-
-		assertThat(writers)
-			.hasSize(3)
-			.allSatisfy(element -> assertThat(element).hasSize(5));
-
-		assertThat(writers[0]).allSatisfy(collection -> assertThat(collection).isEmpty());
-		assertThat(writers[1]).allSatisfy(collection ->
-			assertThat(collection)
-				.hasSize(2)
-				.hasAtLeastOneElementOfType(ConsoleWriter.class)
-				.hasAtLeastOneElementOfType(FileWriter.class)
-		);
-		assertThat(writers[2]).allSatisfy(collection -> assertThat(collection).isEmpty());
-	}
-
-	/**
-	 * Verifies that a writer with a defined severity level, which is above the minimum severity level, will be only
-	 * assigned to the configured severity levels.
-	 */
-	@Test
-	public void levelOfWriterAboveMinimumLevel() {
-		Configuration.set("writer", "console");
-		Configuration.set("writer.level", "WARN");
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.TRACE, false);
-
-		assertThat(writers).hasSize(2).allSatisfy(element -> {
-			assertThat(element).hasSize(5);
-			assertThat(element[Level.TRACE.ordinal()]).isEmpty();
-			assertThat(element[Level.DEBUG.ordinal()]).isEmpty();
-			assertThat(element[Level.INFO.ordinal()]).isEmpty();
-			assertThat(element[Level.WARN.ordinal()])
-				.hasSize(1)
-				.allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-			assertThat(element[Level.ERROR.ordinal()])
-				.hasSize(1)
-				.allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-		});
-	}
-
-	/**
-	 * Verifies that a writer with a defined severity level, which is below the minimum severity level, will be not
-	 * assigned to severity levels below the minimum severity level.
-	 */
-	@Test
-	public void levelOfWriterBelowMinimumLevel() {
-		Configuration.set("writer", "console");
-		Configuration.set("writer.level", "trace");
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.INFO, false);
-
-		assertThat(writers).hasSize(2).allSatisfy(element -> {
-			assertThat(element).hasSize(5);
-			assertThat(element[Level.TRACE.ordinal()]).isEmpty();
-			assertThat(element[Level.DEBUG.ordinal()]).isEmpty();
-			assertThat(element[Level.INFO.ordinal()])
-				.hasSize(1)
-				.allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-			assertThat(element[Level.WARN.ordinal()])
-				.hasSize(1)
-				.allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-			assertThat(element[Level.ERROR.ordinal()])
-				.hasSize(1)
-				.allSatisfy(writer -> assertThat(writer).isInstanceOf(ConsoleWriter.class));
-		});
-	}
-
-	/**
-	 * Verifies that writers with the severity level {@link Level#OFF OFF} will not be created.
-	 */
-	@Test
-	public void levelOfWriterIsOff() {
-		Configuration.set("writer", "console");
-		Configuration.set("writer.level", "off");
-
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.INFO, false);
-
-		assertThat(writers).hasSize(2).allSatisfy(element ->
-			assertThat(element).hasSize(5).allSatisfy(collection -> assertThat(collection).isEmpty())
-		);
-	}
-
-	/**
-	 * Verifies that a {@link ThrowableFilter} can be registered globally.
-	 */
-	@Test
-	public void globalException() {
-		Configuration.set("exception", "drop cause");
-		Configuration.set("writer", "configuration store");
-		
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.INFO, false);
-		ConfigurationStoreWriter writer = (ConfigurationStoreWriter) writers[0][Level.INFO.ordinal()].iterator().next();
-		assertThat(writer.getProperties()).contains(entry("exception", "drop cause"));
-	}
-	
-	/**
-	 * Verifies that a {@link ThrowableFilter} can be registered on a writer.
-	 */
-	@Test
-	public void localException() {
-		Configuration.set("writer", "configuration store");
-		Configuration.set("writer.exception", "unpack");
-		
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.INFO, false);
-		ConfigurationStoreWriter writer = (ConfigurationStoreWriter) writers[0][Level.INFO.ordinal()].iterator().next();
-		assertThat(writer.getProperties()).contains(entry("exception", "unpack"));
-	}
-	
-	/**
-	 * Verifies that a global {@link ThrowableFilter} can be overridden by another one directly registered on a writer.
-	 */
-	@Test
-	public void overrideException() {
-		Configuration.set("exception", "drop cause");
-		Configuration.set("writer", "configuration store");
-		Configuration.set("writer.exception", "unpack");
-		
-		Collection<Writer>[][] writers = ConfigurationParser.createWriters(emptyList(), Level.INFO, false);
-		ConfigurationStoreWriter writer = (ConfigurationStoreWriter) writers[0][Level.INFO.ordinal()].iterator().next();
-		assertThat(writer.getProperties()).contains(entry("exception", "unpack"));
 	}
 
 }
